+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/procs/compare/v0.9.5...Unreleased) - ReleaseDate
 
+* [Changed] --watch and --watch-interval option [#36](https://github.com/dalance/procs/issues/36)
+
 ## [v0.9.5](https://github.com/dalance/procs/compare/v0.9.4...v0.9.5) - 2020-01-30
 
 * [Fixed] Remove --suid to fix security vulnerability (arbitrary command execution by root) [#38](https://github.com/dalance/procs/issues/38)